[nemo-connectivity] Add DataConnection component. Contributes to JB#38786
nemo-connectivity-devel package provides C++ headers for
connectivity handling. Currently it contains only DataConnection
component that can be used to connect and disconnect mobile data
connection. The DataConnection is a composite of connman and ofono
classes that are needed when opening mobile data connection. User
of the component needs to show proper error messages e.g. when
roaming is not allowed.

DataConnection is also explosed to QML through Nemo.Connectivity
import (nemo-qml-plugin-connectivity).

File structure of the project is aligned with Nemo.DBus.