[nemo-qml-plugin-connectivity] Remove some bad API. Fixes JB#45405
connectionSelectorOpened property was added with commit 0d3a6159efbb9
but the implementation was somewhat insufficient, the property
didn't even have NOTIFY signalling. It's not anymore used either
so just getting rid of it now.

haveNetworkConnectivity() is almost the same as online, except that
it's not as usable without change signal and didn't return the correct
value immediately anyway.