[connectivity] Add support for ConnMan tech powered state update. Fixes JB#55593
Implement support for changing the powered state also in ConnMan when
the modem for to be used with cellular is selected. This is done using
the technology API of libconnman-qt and the purpose is to indicate the
proper powered state in ConnMan as well in order to get the value set in
the settings file so ConnMan does not disable mobile data when loading
the values.

Also enforce the powered state from ofono to ConnMan if those differ
according to the signal that is received. Do not do the changes when
doing the initial setup as the powered state might not have been set yet
when initializing technology via libconnman-qt. This relies on the
changes coming from ofono.