[connectivity] Do not connect to poweredChanged signal due to race. Fixes JB#56031 OMP#JOLLA-460
There may and usually is a race if the Cellular technology is disabled
via ConnMan D-Bus API as the signal that is disabled with this change
would be catched here first before the state is updated via Ofono and
thus, ConnMan will be sent a message to restore the old state. The
management of the state differences must be improved here or perhaps
better yet, to be done completely at lower levels. Thus, there are now
TODO comments left and the signal connect/disconnect are now only
commented out as a reminders.